Remediate handoff §7/§8/§9/§10; make scripts/ chi-aware
Closes the four audited tracks. Tests 124 -> 135, all green.

§7 bugs: 7.1-7.11 fixed (7.12 for scripts/ only). Highlights: fork_rng
around the Fisher seeds; cfg.n_obs ends the two-observation-length split
that silently dropped all five overlay figures; reproducible priors;
FWHM walks outward from the argmax; implicit_euler deleted; non-uniform
dt and off-PSD-grid interpolation now refused rather than faked.

§8 perf: calibration batching decoupled (defaults bit-identical);
memory fixes in gen_training_data/decorrelate/psd_welch/campaigns;
Nadrowski noise hoisted (eager + JIT); cycle_average and _group_g
rewritten; verified bit-identical or to float32 epsilon.

§9: fixed the shadowed `labels` module (its consumer was already
mis-wired), collapsed four duplicate row widgets, documented every panel
class, replaced 12 stale file:line citations with function names.

§10 GUI: L1-L18. Splitter persistence, form growth policy, field minimum
widths, no 460px cap, adaptive stacks, results splitter, OS text scaling.

scripts/: all 11 routed through new scripts/_common.py. sbc_characterize,
retrain_convergence and degeneracy_map are chi-aware; the three whose
metrics are defined over the single-frequency feature set now refuse
CHI=1 rather than answer wrongly. Posteriors always write a mode sidecar,
and a cross-mode load fails before the simulation spend.

BREAKING: B1_log_Q changed (7.6). Do not re-evaluate a pre-fix posterior
(incl. posterior_07012026) with post-fix statistics.

Seven audit recommendations were rejected after verification and are
corrected in place in PRISM_HANDOFF.md rather than dropped.

+++ core/FDT/spectral.py
@@ -24,7 +24,11 @@ def psd_welch(x: torch.Tensor, dt: float, nperseg: int, overlap: float = 0.5) ->
     :param overlap: fractional segment overlap. SciPy default is 0.5.
     :return: (omegas, G) -- omegas shape (nperseg//2 + 1,), G one-sided PSD shape (nperseg//2 + 1,).
     """
-    x = x.to(torch.float64)
+    # NOT `x = x.to(torch.float64)`. Promoting the WHOLE ensemble up front doubles a tensor that is
+    # ~1.6 GB at Campaign-1 defaults, when only the (M, nperseg) slice inside the loop is ever
+    # arithmetic -- ~33 MB. FDT runs on host RAM (cpu_device()), so that was 1.6 GB of resident
+    # working set for a diagnostic. The promotion moves onto `seg` below; every accumulation,
+    # the window and the output stay float64, so the result is unchanged.
     M, T_total = x.shape
     step = max(1, int(nperseg * (1 - overlap)))
     n_segs = 1 + (T_total - nperseg) // step
@@ -36,7 +40,7 @@ def psd_welch(x: torch.Tensor, dt: float, nperseg: int, overlap: float = 0.5) ->
 
     psd_accum = torch.zeros(nperseg // 2 + 1, dtype=torch.float64, device=x.device)
     for i in range(n_segs):
-        seg = x[:, i * step : i * step + nperseg]
+        seg = x[:, i * step : i * step + nperseg].to(torch.float64)   # promote the SLICE, not x
         seg = seg - seg.mean(dim=-1, keepdim=True)   # detrend each segment
         seg = seg * window
         X = torch.fft.rfft(seg, dim=-1)              # (M, nperseg//2 + 1)
